refactor: Move game result, round result object to DTO
