+++ src/main/java/racingcar/view/ResultView.java
@@ -1,9 +1,10 @@
 package racingcar.view;
 
 import racingcar.dto.CarResultDto;
+import racingcar.dto.GameResultDto;
+import racingcar.dto.RoundResultDto;
 
 import java.util.List;
-import java.util.function.Consumer;
 
 public class ResultView {
     private static final String RESULT_DELIMITER = "-";
@@ -12,9 +13,9 @@ public class ResultView {
 
     private static final String WINNER_DELIMITER = ", ";
 
-    public static void printGameResult(List<List<CarResultDto>> gameResult) {
+    public static void printGameResult(GameResultDto gameResult) {
         printResultComment();
-        gameResult.forEach(printRoundResult());
+        gameResult.getRoundResults().forEach(ResultView::printRoundResult);
     }
 
     public static void printWinners(List<String> winners) {
@@ -22,18 +23,14 @@ public class ResultView {
         System.out.println(winnersString + "가 최종 우승했습니다.");
     }
 
-    private static Consumer<List<CarResultDto>> printRoundResult() {
-        return eachRound -> {
-            eachRound.forEach(printCarResult());
-            System.out.println();
-        };
+    private static void printRoundResult(RoundResultDto roundResult) {
+        roundResult.getResults().forEach(ResultView::printCarResult);
+        System.out.println();
     }
 
-    private static Consumer<CarResultDto> printCarResult() {
-        return carResult -> {
-            printDelimitedResult(carResult.position, carResult.name);
-            System.out.println();
-        };
+    private static void printCarResult(CarResultDto carResult) {
+        printDelimitedResult(carResult.position, carResult.name);
+        System.out.println();
     }
 
     private static void printDelimitedResult(Integer position, String name) {
